requirements.txt
Django==1.10.6
django-jchart==0.2.0
